共通Descriptionとして切り出す。
